+++ backend/src/main/java/ru/edjll/backend/service/PostService.java
@@ -1,10 +1,13 @@
 package ru.edjll.backend.service;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import ru.edjll.backend.dto.PostDto;
 import ru.edjll.backend.entity.Post;
 import ru.edjll.backend.repository.PostRepository;
 
+import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.Collection;
 
@@ -17,19 +20,28 @@ public class PostService {
         this.postRepository = postRepository;
     }
 
-    public PostDto save(Post post) {
+    public PostDto save(Post post, Principal principal) {
+        if (post.getUser() == null || !principal.getName().equals(post.getUser().getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Don't have rights");
+        }
         post.setCreatedDate(LocalDateTime.now());
         Post savedPost = postRepository.save(post);
         return postRepository.getPostDtoById(savedPost.getId());
     }
 
-    public PostDto edit(Post post) {
+    public PostDto edit(Post post, Principal principal) {
+        if (post.getUser() == null || !principal.getName().equals(post.getUser().getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Don't have rights");
+        }
         post.setModifiedDate(LocalDateTime.now());
         Post savedPost = postRepository.save(post);
         return postRepository.getPostDtoById(savedPost.getId());
     }
 
-    public void delete(Post post) {
+    public void delete(Post post, Principal principal) {
+        if (post.getUser() == null || !principal.getName().equals(post.getUser().getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Don't have rights");
+        }
         postRepository.deleteById(post.getId());
     }
